moved RegisterUserWarnings, RegisterUser to account.ts
